+++ packages/core/src/markdown/index.ts
@@ -1,0 +1,2 @@
+export { parseMarkdownToIR } from './parse';
+export { serializeIRToMarkdown } from './serialize';